requirements.txt
pandas
numpy
requests
pvlib
xgboost
scikit-learn
joblib
matplotlib

# Web / API (added by teammate)
fastapi
uvicorn[standard]
pydantic>=2.0

# Edge Data Layer
paho-mqtt>=1.6.1
sqlalchemy>=2.0.0

# Orchestrator FSM
transitions>=0.9.0

# Testing
pytest>=8.0.0

# Dev-only Broker
amqtt>=0.11.0

# Strategic LLM Agent
google-genai>=0.1.0
